fix validation services

+++ src/services/course.service.ts
@@ -1,7 +1,8 @@
 import { BaseCrudService } from "./base.service";
 import { courseRepository } from "./instance.repo";
 import { courses } from "../db/schema/course.db";
-import type { CourseDetailsInsert } from "../db/index.types";
+import type { CourseDetailsInsert } from "../db/schema/course.db";
+import { NotFoundError } from "../errors/not-found.error";
 
 export class Course extends BaseCrudService<typeof courses> {
   constructor() {
@@ -9,21 +10,43 @@ export class Course extends BaseCrudService<typeof courses> {
   }
 
   async getDetailById(detailId: number) {
-    return courseRepository.getDetailById(detailId);
+    const record = await courseRepository.getDetailById(detailId);
+    if (!record) {
+      throw new NotFoundError(
+        `cannot get: detail ${this.primaryKey} ${detailId} not found`,
+      );
+    }
+    return record;
   }
 
   async addDetails(courseId: number, newCourseDetail: CourseDetailsInsert) {
-    return courseRepository.addDetails(courseId, newCourseDetail);
+    const record = await courseRepository.addDetails(courseId, newCourseDetail);
+    if (!record) {
+      throw new Error("failed to create the record.");
+    }
+    return record;
   }
 
   async updateDetails(
     detailId: number,
     newDetail: Partial<CourseDetailsInsert>,
   ) {
+    const exists = await this.getDetailById(detailId);
+    if (!exists) {
+      throw new NotFoundError(
+        `cannot update: detail ${this.primaryKey} ${detailId} not found`,
+      );
+    }
     return courseRepository.updateDetails(detailId, newDetail);
   }
 
   override async delete(id: number) {
+    const exists = await this.getDetailById(id);
+    if (!exists) {
+      throw new NotFoundError(
+        `cannot delete: detail ${this.primaryKey} ${id} not found`,
+      );
+    }
     return courseRepository.deleteCourseWithDetails(id);
   }
 }
