+++ src/services/project.service.ts
@@ -1,7 +1,8 @@
 import { BaseCrudService } from "./base.service";
 import { projectRepository } from "./instance.repo";
 import { projects } from "../db/schema/project.db";
-import type { ProjectDetailsInsert } from "../db/index.types";
+import type { ProjectDetailsInsert } from "../db/schema/project.db";
+import { NotFoundError } from "../errors/not-found.error";
 
 export class Project extends BaseCrudService<typeof projects> {
   constructor() {
@@ -9,21 +10,46 @@ export class Project extends BaseCrudService<typeof projects> {
   }
 
   async getDetailById(detailId: number) {
-    return projectRepository.getDetailById(detailId);
+    const record = await projectRepository.getDetailById(detailId);
+    if (!record) {
+      throw new NotFoundError(
+        `cannot get: detail ${this.primaryKey} ${detailId} not found`,
+      );
+    }
+    return record;
   }
 
   async addDetails(projectId: number, newProjectDetail: ProjectDetailsInsert) {
-    return projectRepository.addDetails(projectId, newProjectDetail);
+    const record = await projectRepository.addDetails(
+      projectId,
+      newProjectDetail,
+    );
+    if (!record) {
+      throw new Error("failed to create the record.");
+    }
+    return record;
   }
 
   async updateDetails(
     detailId: number,
     newDetail: Partial<ProjectDetailsInsert>,
   ) {
-    return await projectRepository.updateDetails(detailId, newDetail);
+    const exists = await this.getDetailById(detailId);
+    if (!exists) {
+      throw new NotFoundError(
+        `cannot update: detail ${this.primaryKey} ${detailId} not found`,
+      );
+    }
+    return projectRepository.updateDetails(detailId, newDetail);
   }
 
   override async delete(id: number) {
+    const exists = await this.getDetailById(id);
+    if (!exists) {
+      throw new NotFoundError(
+        `cannot delete: detail ${this.primaryKey} ${id} not found`,
+      );
+    }
     return projectRepository.deleteProjectWithDetails(id);
   }
 }
